# Optimize timeline generation