Those are somewhat low-hanging fruits for timeline generation.

## Before

Before the changes, generating my Home and two (small) list timelines from scratch took that amount of time:
```
 61.001444   6.975359  67.976803 (176.038725)
 11.526668   1.181717  12.708385 ( 33.958586)
  3.576163   0.401200   3.977363 ( 15.360795)
```

And when the timelines were already generated, filling them (which would be a no-op) took that amount of time:
```
 51.031345   6.324234  57.355579 (169.474034)
 10.023077   1.142712  11.165789 ( 23.435064)
  2.929509   0.277567   3.207076 ( 10.042510)
```

## After

After this PR's changes, generating the same timelines from scratch takes a reduced amount of time:
```
  9.817145   1.210652  11.027797 ( 19.110152)
  5.465910   0.680844   6.146754 ( 11.889094)
  2.949545   0.423947   3.373492 (  9.529031)
```

And when already generated:
```
  5.764612   0.619574   6.384186 ( 10.486542)
  2.226865   0.267550   2.494415 (  4.391762)
  1.532688   0.127098   1.659786 (  2.895375)
```